+++ lib/gql/relay/alias-deprecated-fields.decorator.d.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata';
-import { RelayFilterBase } from './relay.interface';
+import { RelayFilterBase, RelayInputTypeBase } from './relay.interface';
 import { FieldOptions } from 'type-graphql';
 /**
  *
@@ -15,15 +15,19 @@ import { FieldOptions } from 'type-graphql';
  * defined with getters/setters in the base-attribs file:
  * `{ foo: get(): string|undefined }` and `{ foo?: string }` are treated very differently.
  */
-export declare function AliasFromDeprecatedField<T, K = keyof T>(deprecated_prop_name: K, field_options: FieldOptions): PropertyDecorator;
+export declare function AliasFromDeprecatedField<T, K = keyof T>(deprecated_prop_name: K, field_options?: FieldOptions): PropertyDecorator;
 /**
  *
- * @param filter can be Filter|Input|Update
+ * @param filterOrInputType can be Filter|Input|Update
  * CANNOT BE true true relay node object, which are very special.
+ *
+ * @throws Error if the passed object is a NodeRelay
+ * (discriminated by presence of Symbol(modelKey)`)
  *
  * @description Remove any deprecated values from the object before it gets turned into an
  * instruction to sequelize for create/update/filter.
  *
- * @returns A **clone** of the original filter/input with the deprecated values removed
+ * @returns If any args are marked as deprecated, a **clone** of the original filter/input
+ *  with the deprecated values removed, otherwise: the original object
  */
-export declare function cloneAndTransposeDeprecatedValues<T extends RelayFilterBase<any>>(filter: T): T;
+export declare function cloneAndTransposeDeprecatedValues<T extends RelayFilterBase<any> | RelayInputTypeBase<any>>(filterOrInputType: T): T;